opt(tree-selector): cascader
